+++ src/main/java/com/mystdev/recicropal/content/crop/bottle_gourd/BottleGourdItem.java
@@ -152,6 +152,6 @@ public class BottleGourdItem extends BlockItem {
 
     @Override
     public @Nullable ICapabilityProvider initCapabilities(ItemStack stack, @Nullable CompoundTag nbt) {
-        return new FluidHandlerItemStack(stack, BottleGourdTank.CAPACITY);
+        return new FluidHandlerItemStack(stack, BottleGourdTank.configuredCapacity());
     }
 }